Deprecate configuration file in favor of environment variables
This change removes the need for configuration file, by opting for an
environment variable to store the API token. Since there is only one variable
needed, using an environment variable simplifies the process by avoiding the
need to look up and find the configuration file, parsing and validation.

This commit should also resolve #4.